feat(all): new component to general profile data
